requirements.txt
dlib
face_recognition
numpy
pandas
opencv-python
pillow